Make error Toast a nicer Snackbar, refresh theme on change

+++ app/src/main/java/fr/outadoc/quickhass/feature/grid/vm/EntityGridViewModel.kt
@@ -36,7 +36,7 @@ class EntityGridViewModel(
     private val _loadingEntityIds = MutableLiveData<Set<@StringEntityId String>>()
     val loadingEntityIds: LiveData<Set<@StringEntityId String>> = _loadingEntityIds
 
-    val refreshIntervalSeconds: Int
+    val refreshIntervalSeconds: Long
         get() = prefs.refreshIntervalSeconds
 
     fun loadShortcuts() {
